feat: add pagination   add node_modules in ignorePatterns in the eslintrc   add tailwind.config   remove yarn.lock   moves TDataDB to db config   moves saveAssociatedProductInDB to models
