+++ src/config/db.ts
@@ -1,6 +1,6 @@
-import type { TDataDB } from '@/db/types'
 import { MongoClient } from 'mongodb'
 import dotenv from 'dotenv'
+import type { TData } from '@/bot/types'
 
 dotenv.config()
 
@@ -9,6 +9,10 @@ const uri = process.env.MONGO_URI || ''
 const client = new MongoClient(uri)
 
 const database = client.db('baratao')
+
+export type TDataDB = {
+  createdAt: string
+} & TData
 
 export const associatedCollection = database.collection<TDataDB>('associated')
