--- src/db/index.ts
@@ -1,2 +1,0 @@
-export * from './mutations'
-export * from './types'
