--- src/db/mutations.ts
@@ -1,6 +1,0 @@
-import type { TDataDB } from './types'
-import { associatedCollection } from '@/config'
-
-export const saveAssociatedProductInDB = async (product: TDataDB) => {
-  await associatedCollection.insertOne(product)
-}
