--- src/db/types.ts
@@ -1,5 +1,0 @@
-import type { TData } from '@/bot/types'
-
-export type TDataDB = {
-  createdAt: string
-} & TData
