+++ src/models/index.ts
@@ -1,13 +1,16 @@
-import { associatedCollection } from '@/config'
-import { type TDataDB } from '@/db'
+import { type TDataDB, associatedCollection } from '@/config'
 
-export const getAssociatedProducts = async (page: number) => {
+export const saveAssociatedProductInDB = async (product: TDataDB) => {
+  await associatedCollection.insertOne(product)
+}
+
+export const getAssociatedProducts = async (page: number, limit: number) => {
   const total = await associatedCollection.countDocuments()
   const cursor = associatedCollection
     .find()
     .sort({ createdAt: -1 })
     .skip(100 * (page - 1))
-    .limit(100)
+    .limit(limit)
   const products: TDataDB[] = []
   for await (const product of cursor) {
     products.push(product)
