+++ tailwind.config.js
@@ -1,0 +1,8 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: ['./src/**/*.{pug,js}'],
+  theme: {
+    extend: {}
+  },
+  plugins: [require('tailwind-scrollbar')]
+}
